requirements.txt
fastapi
uvicorn
python-dotenv
motor
pydantic[email]
PyJWT
passlib[bcrypt]
langchain
langchain-google-genai
langchain-community
pypdf
python-multipart
cors
